move custom recurrence modal into recurrence options
